New and shiny message implementation.
Now we have system scoped messages, entity scoped messages and field scoped messages. Each
one may be an error, an information or a warning. Each scope is handled by page tag
so there is no need to do it individually. Validators return a list of this messages.

This may need some fixes but is way better than previous system.